Task5 - Tweaks on App.py, rquirements.txt and log module
