+++ client_data/.part-00000-74bbc4ee-cdcf-4d39-8495-38f940af07eb-c000.csv.crc
