+++ requirements.txt
@@ -1,0 +1,2 @@
+py4j==0.10.9.7
+pyspark==3.5.0
